Refactor into appropriate package.
